requirements.txt
mutagen-1.47.0
PyQt6